requirements.txt
asgiref==3.10.0
click==8.3.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
pillow==12.0.0
psycopg2==2.9.11
sqlparse==0.5.3
uvicorn==0.37.0
whitenoise==6.11.0
